Add Readably to compatible apps
Hi There!

I'm Isaias, I am the developer of Readably, an rss reader for Android. Readably has recently picked up Fever API support, so I want to ask if you could add it to the list of Fever API compatible apps in this file. Here is the link:

https://play.google.com/store/apps/details?id=com.isaiasmatewos.readably

Thanks, Isaias